+++ vite.plugin.config.ts
@@ -53,19 +53,81 @@ function onlyBuildOptions(mode: string) {
       url: 'https://www.staticfile.org/',
       // modules 顺序 vue, vue-demi 必须保持当前顺序加载，否则会出现加载错误问题
       modules: [
-        'vue',
-        'vue-demi',
-        'pinia',
-        'naive-ui',
-        'vue-router',
-        'vue-i18n',
-        'dayjs',
-        'echarts',
-        'xlsx',
-        'axios',
-        'print-js',
-        'clipboard',
-        'lodash-es',
+        {
+          name: 'vue',
+          global: 'Vue',
+          resolve: 'https://cdn.staticfile.org/vue/3.3.11/vue.global.min.js',
+        },
+        {
+          name: 'vue-demi',
+          global: 'VueDemi',
+          resolve:
+            'https://cdn.staticfile.org/vue-demi/0.14.6/index.iife.min.js',
+        },
+        {
+          name: 'naive-ui',
+          global: 'naive',
+          resolve: 'https://cdn.staticfile.org/naive-ui/2.35.0/index.min.js',
+        },
+        {
+          name: 'pinia',
+          global: 'Pinia',
+          resolve: 'https://cdn.staticfile.org/pinia/2.1.7/pinia.iife.min.js',
+        },
+        {
+          name: 'vue-router',
+          global: 'VueRouter',
+          resolve:
+            'https://cdn.staticfile.org/vue-router/4.2.5/vue-router.global.min.js',
+        },
+        {
+          name: 'vue-i18n',
+          global: 'VueI18n',
+          resolve:
+            'https://cdn.staticfile.org/vue-i18n/9.8.0/vue-i18n.global.min.js',
+        },
+        {
+          name: 'dayjs',
+          global: 'dayjs',
+          resolve: 'https://cdn.staticfile.org/dayjs/1.11.10/dayjs.min.js',
+        },
+        {
+          name: 'echarts',
+          global: 'echarts',
+          resolve: 'https://cdn.staticfile.org/echarts/5.4.3/echarts.min.js',
+        },
+        {
+          name: 'xlsx',
+          global: 'XLSX',
+          resolve: 'https://cdn.staticfile.org/xlsx/0.18.5/xlsx.full.min.js',
+        },
+        {
+          name: 'axios',
+          global: 'axios',
+          resolve: 'https://cdn.staticfile.org/axios/1.6.2/axios.min.js',
+        },
+        {
+          name: 'print-js',
+          global: 'printJS',
+          resolve: 'https://cdn.staticfile.org/print-js/1.6.0/print.min.js',
+        },
+        {
+          name: 'dom-to-image',
+          global: 'domtoimage',
+          resolve:
+            'https://cdn.staticfile.org/dom-to-image/2.6.0/dom-to-image.min.js',
+        },
+        {
+          name: 'clipboard',
+          global: 'ClipboardJS',
+          resolve:
+            'https://cdn.staticfile.org/clipboard.js/2.0.11/clipboard.min.js',
+        },
+        {
+          name: 'lodash-es',
+          global: '_',
+          resolve: 'https://cdn.staticfile.org/lodash.js/4.17.21/lodash.min.js',
+        },
       ],
     }),
   ]